refactor(stream): rename ChunkStream to DistributionStream
