+++ libraries/stream-extra/src/distribution.ts
