+++ libraries/stream-extra/src/index.ts
@@ -1,7 +1,7 @@
 export * from "./buffered-transform.js";
 export * from "./buffered.js";
-export * from "./chunk.js";
 export * from "./decode-utf8.js";
+export * from "./distribution.js";
 export * from "./duplex.js";
 export * from "./gather-string.js";
 export * from "./inspect.js";
